nvcodec: Fix broken ABI in cuda stub header to fix nvenc with opengl
Fix the broken ABI introduced by the commit 367e742e5dd53400d212ce07d0ac0745f3535ac3
From CUDA Toolkit 3.2, size_t has been used in CUDA_MEMCPY2D structure
instead of unsigned int.

+++ sys/nvcodec/stub/cuda.h
@@ -64,7 +64,7 @@ typedef struct
   gconstpointer srcHost;
   CUdeviceptr srcDevice;
   CUarray srcArray;
-  guint srcPitch;
+  gsize srcPitch;
 
   gsize dstXInBytes;
   gsize dstY;
